requirements.txt
streamlit>=1.38.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.2
lightgbm>=4.5.0
shap>=0.46.0
joblib>=1.4.2
langchain>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
matplotlib>=3.9.2
plotly>=5.23.0
python-dotenv>=1.0.1
tqdm>=4.66.5
